+++ scripts/module.js
@@ -1,4 +1,4 @@
-import { setupAPI } from "./helpers.js";
+import { getSetting, setupAPI } from "./helpers.js";
 import { registerKeybindings, registerSettings } from "./settings.js";
 
 Hooks.once("init", async function () {
@@ -7,4 +7,19 @@ Hooks.once("init", async function () {
   setupAPI();
 });
 
-Hooks.once("ready", async function () {});
+Hooks.once("ready", async function () {
+  Hooks.on("createChatMessage", async function (msg, _status, userid) {
+    if (game.user.id !== userid) return;
+    if (
+      msg?.item?.system?.actionType?.value === "reaction" &&
+      !msg?.flags?.pf2e?.context?.type &&
+      getSetting("reaction.enabled")
+    ) {
+      const tokDoc = msg?.token || msg?.actor?.prototypeToken;
+      const characterArt = tokDoc?.ring?.enabled
+        ? tokDoc?.ring?.subject?.texture || tokDoc?.texture?.src
+        : tokDoc?.texture?.src;
+      game.objection.api.objection({ type: "objection", characterArt });
+    }
+  });
+});